+++ src/types/form-element.ts
@@ -22,7 +22,6 @@ export const enum FormElementType {
   NUMBER = 'number',
   PASSWORD = 'password',
   RADIO = 'radio',
-  SECRET = 'secret',
   SELECT = 'select',
   SLIDER = 'slider',
   STRING = 'string',
@@ -552,18 +551,6 @@ export interface CustomButtonOptions {
 }
 
 /**
- * Secret Options
- */
-export interface SecretOptions {
-  /**
-   * Value
-   *
-   * @type {string}
-   */
-  value: string;
-}
-
-/**
  * Link Target
  */
 export enum LinkTarget {
@@ -628,7 +615,6 @@ export type FormElement = FormElementBase &
     | ({ type: FormElementType.PASSWORD } & PasswordOptions)
     | ({ type: FormElementType.DATETIME } & DateTimeOptions)
     | ({ type: FormElementType.DATE } & DateOptions)
-    | ({ type: FormElementType.SECRET } & SecretOptions)
     | ({ type: FormElementType.BOOLEAN } & BooleanOptions)
     | ({ type: FormElementType.BUTTON } & CustomButtonOptions)
     | ({ type: FormElementType.FILE } & FileOptions)