+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added isEscaping option for TextArea and CodeEditor elements (#581)
 - Added support streaming state for data (#577)
+- Refactoring form elements (#576)
 
 ## 5.0.0 (2024-12-23)